Add builder with description and no arguments
