+++ Packages/UGF.Description/Runtime/DescribedBuilderAsset.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 236c483725874e09ae89dcab4231acdc
-timeCreated: 1606680910
+guid: d2d179049e3e4529ad7e2471ce3079ac
+timeCreated: 1606768430
