+++ Packages/UGF.Description/Runtime/DescribedBuilderAsset`2.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 236c483725874e09ae89dcab4231acdc
+timeCreated: 1606680910
