+++ Packages/UGF.Description/Runtime/DescribedWithDescriptionBuilderAsset.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 1a1a72a18b374379a7795b4061b2c205
-timeCreated: 1606681097
+guid: 15b75c6406d74fb9b84e2f16dace6ff4
+timeCreated: 1606768603
