+++ Packages/UGF.Description/Runtime/DescribedWithDescriptionBuilderAsset`1.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1a1a72a18b374379a7795b4061b2c205
+timeCreated: 1606681097
