+++ Packages/UGF.Description/Runtime/IDescribedBuilder.cs
@@ -2,7 +2,7 @@
 
 namespace UGF.Description.Runtime
 {
-    public interface IDescribedBuilder<in TArguments> : IBuilder<TArguments, IDescribed>
+    public interface IDescribedBuilder : IBuilder<IDescribed>
     {
     }
 }
