+++ Packages/UGF.Description/Runtime/IDescribedBuilder.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: f9f676fc33ef4873b8d465aa21d26cbf
-timeCreated: 1606678129
+guid: 0db0e93e525e49c5a37d482ec1a5876e
+timeCreated: 1606768510
