+++ Packages/UGF.Description/Runtime/IDescribedBuilder`1.cs
@@ -1,0 +1,8 @@
+﻿using UGF.Builder.Runtime;
+
+namespace UGF.Description.Runtime
+{
+    public interface IDescribedBuilder<in TArguments> : IBuilder<TArguments, IDescribed>
+    {
+    }
+}
