+++ Packages/UGF.Description/Runtime/IDescribedBuilder`1.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f9f676fc33ef4873b8d465aa21d26cbf
+timeCreated: 1606678129
